'prototype' undefined fix for dd-gridstack.js
* fix #2039
* moved all the D7D gridstack code into the gridstack.ts file as I gave up on the idea to build s a smaller static only code and delay load the D&D (which is much much smaller now)
as it was causing some build issues for some users...
